Move ReadTest in its own package, make some test parameterized
